+++ public/js/main.js
@@ -1,0 +1,3 @@
+document.addEventListener('DOMContentLoaded', function (event) {
+  window.anchors.add()
+})