+++ src/constant.js
@@ -7,7 +7,8 @@ const {
   CACHE_TTL = TWENTY_FOUR_HOURS,
   LOG_LEVEL = 'tiny',
   AVATAR_SIZE = 400,
-  AVATAR_TIMEOUT = 60000
+  AVATAR_TIMEOUT = 60000,
+  ALLOWED_REQ_HEADERS = ['user-agent', 'accept', 'x-api-key']
 } = process.env
 
 module.exports = {
@@ -16,5 +17,6 @@ module.exports = {
   CACHE_TTL,
   LOG_LEVEL,
   AVATAR_SIZE,
-  AVATAR_TIMEOUT
+  AVATAR_TIMEOUT,
+  ALLOWED_REQ_HEADERS
 }